Enable dumb saml IdP for use with SP
Modify quickstart configuration for idp-with-metadata to provide an
identity provider for use with sp-with-metadata. The IdP works without
signature verification or encryption, so is unsuitable for use in prod
but can be used as a starting point for further investigation, or for
testing to ensure a SAML login path is working.

This IdP requires an idp security domain to be defined in wildfly,
there is one available that is distributed separately from this git
repo.